修改baseModel结构体uid的类型

+++ app/models/model.go
@@ -8,7 +8,7 @@ import (
 
 // BaseModel 模型基类
 type BaseModel struct {
-	ID uint64 `gorm:"column:id;primaryKey;autoIncrement;" json:"id,omitempty"`
+	ID int `gorm:"column:id;primaryKey;autoIncrement;" json:"id,omitempty"`
 }
 
 // CommonTimestampsField 时间戳
